Cleaning up stutters while walking by only calling keyboard api when walking state changes.  Walking is now very smooth.
Also trying to deal with the jitter in reading from the reed sensor (hence the attempt to apply debounce code which isn't working).  Every time user just pedals once, there is a false calculation of a walk speed which triggers a brief press of the "w" key (more than a single press, so can't count this as implementing Feature #162).